use lsblk to attempt to guess partition's fs type
If not found it will fallback to use other tools.

Note that fs_type function is not expected to return error code
different than zero even if it failed to detect type.

If filesystem is undermined then "NOT-DETECTED" will be echoed.

This change will prevent mount to hang on extended part (#768902)

Bug-Debian: 768902
Origin: https://github.com/rzr/os-prober/tree/sandbox/rzr/for-upstream
Signed-off-by: Philippe Coval <rzr@gna.org>